refactor: Consolidating initial Checkers and raising questions
